[AIRLFOW-YYY] Move scheduler command to separate file
